requirements.txt
streamlit
opencv-python-headless
numpy
mtcnn
tensorflow
Pillow